fix build for 64bit architectures
svn path=/trunk/kdelibs/; revision=257742

+++ khtml/misc/loader_jpeg.cpp
@@ -279,7 +279,7 @@ int KJPEGFormat::decode(QImage& image, QImageConsumer* consumer, const uchar* bu
         qDebug("valid_buffer_len %d", jsrc.valid_buffer_len);
         qDebug("skip_input_bytes %d", jsrc.skip_input_bytes);
 #endif
-        int skipbytes = kMin((unsigned) jsrc.valid_buffer_len, jsrc.skip_input_bytes);
+        int skipbytes = kMin((size_t) jsrc.valid_buffer_len, jsrc.skip_input_bytes);
 
         if(skipbytes < jsrc.valid_buffer_len)
             memmove(jsrc.buffer, jsrc.buffer+skipbytes, jsrc.valid_buffer_len - skipbytes);
